fix: Change applicationId type from Long to Integer

+++ src/main/java/com/scms/app/controller/PortfolioItemController.java
@@ -189,7 +189,7 @@ public class PortfolioItemController {
     @PostMapping("/from-program/{applicationId}")
     public ResponseEntity<?> createFromProgramApplication(
             @PathVariable Long portfolioId,
-            @PathVariable Long applicationId,
+            @PathVariable Integer applicationId,
             HttpSession session) {
 
         Integer userId = (Integer) session.getAttribute("userId");
